Fixing unit tests and simplifying primitive array converters.
Moved tests to the appropriate packages.

Changed primitive array converters to use a specific converter - there's
no point in searching the registry when the end type is known.

+++ xll-core/src/main/java/com/mcleodmoores/xl4j/typeconvert/converters/PrimitiveByteXLNumberTypeConverter.java
@@ -29,6 +29,6 @@ public final class PrimitiveByteXLNumberTypeConverter extends AbstractTypeConver
   @Override
   public Object toJavaObject(final Type expectedType, final Object from) {
     ArgumentChecker.notNull(from, "from");
-    return (int) ((XLNumber) from).getValue();
+    return (byte) ((XLNumber) from).getValue();
   }
 }
